// Codexec-IDE/src/api-client/api.ts
export interface LanguageVersion {
    /**
     * Reference to the parent language
     * @type {number}
     * @memberof LanguageVersion
     */
    'language_id': number;
    /**
     * Version identifier of the language
     * @type {string}
     * @memberof LanguageVersion
     */
    'version': string;
    /**
     * Name of the Nix package
     * @type {string}
     * @memberof LanguageVersion
     */
    'nix_package_name': string;
    /**
     * 
     * @type {string}
     * @memberof LanguageVersion
     */
    'template': string;
    /**
     * Search query string
     * @type {string}
     * @memberof LanguageVersion
     */
    'search_query': string;
    /**
     * Whether this is the default version of the language
     * @type {boolean}
     * @memberof LanguageVersion
     */
    'default_version': boolean;
}